Update patch set 2
Patch Set 2:

Well, the file is in n80xx-common, so yeah, currently it's the same. I don't know what marco9333's research[1] based on the files he pulled from a stock n8000 ROM might result in, so it's possible that they could be different, but it might end up being the case that the file we're using is actually wrong for both (I think it comes from i9300).

[1] https://gerrit.omnirom.org/#/c/5416/ - haven't had a chance to actually look into this at all yet.

Patch-set: 2