Update patch set 2
Patch Set 2:

I've already merged this patch (patch set 1) in my github and everything is working as it should be(excluding HDMI I can't test) so far. I'll try to cherry-pick patchset 2 and make the required changes in n8000's tree. I'll report. I don't know what exactly is that loopback device, I've simply ignored it even if it should be a device which route output audio to input.
I've downloaded the datasheet of our audio chip WM8994 but since we don't know how it's wired internally it's quite useless. It's written about two loopbacks devices internally: the first one is the loopback device connected to AIF1 interface but it needs to be switched on with a bit in the chip's internal register at address 0301h, this address is missing in the driver sources header in the kernel so we can't access that register and this functionality is unavailable.

Patch-set: 2