Update patch set 2
Patch Set 2: -Code-Review

The only problem I've found is that on first call after boot there is no audio, but I've read an old commit about reverting a workaround for the same problem in another samsung note phone. Themorpheus is including this fix in his builds and nobody is complaining on xda. About n801x I don't know, someone need to do a deep test,maybe you can create an experimental build with this and the other fixes on my github about audio(I'll push those for review when we'll be 100% sure about this patch which is the most important one), and let the users test and report if something is wrong.

Patch-set: 2