Update patch set 1
Patch Set 1:

I've submitted this as two changes, the first being a direct copy of sbrissen's p4notelte patch, and the second being one which makes it actually work on n80xx. This was just so that it's easier to tell what needed to be changed for n80xx. If I should just squash them into one please let me know.

Also, not yet going to +1 this as I need to test it again, it's been a while since I tested this properly and last time camcorder would crash but I realized that it was only crashing because it defaulted to 1080p which isn't actually supported.

Patch-set: 1