Update patch set 4
Patch Set 4:

Yes, without  https://gerrit.omnirom.org/#/c/6094/ the files in n80xx-common which are in smdk4412-common too, are discarded. This is related directly to the open source camera but 6094 is independent from the camera. Ok il move 6094 as a dependence for this so I can check if I've understood what you said about branches

Patch-set: 4
Label: Code-Review=0