Update patch set 4
Patch Set 4:

By the way, I actually have a keyboard dock now, so I can test the dock audio change. Although right now it seems like the dock drivers are very broken so it may not help, not sure. We'll see when this build finishes (building on my laptop though so it'll take a while).

Patch-set: 4