Update patch set 4
Patch Set 4:

By the way, I have a keyboard dock now, so I can test the dock audio stuff. (Docks actually have several issues right now and basically are broken, so I'm not sure if this is enough to fix that, but we'll see what happens when this build finishes.

Patch-set: 4