Update patch set 4
Patch Set 4:

Dock audio settings are from samsung original files extracted from a stock rom.
The other changes in this patch are working so we need to resolve the "dock problem". I've tried to read the dock driver as well as 30-pin driver with no luck. If it'll take too much I'll divide this into two separate commits to isolate the dock stuff. A user on xda said that with this the audio is reproduced from speakers when on dock (in advancedsettings there's a setting about audio dock but I doubt it's working) without the tablet is mute

Patch-set: 4