Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

I'm just going to go ahead and "ninja-merge" this, but just wanted to let you guys know. It turns out Marco was right and the TARGET_RECOVERY_INITRC was actually used as of about a month ago[1] so the init.rc file that I had originally pulled from the out of data teamwin repo for p4notewifi was actually being used, which caused problems because it was so out of date. For instance, adb would not work correctly in recovery. Removing this and going back to the default init.rc fixes this.

[1] https://github.com/omnirom/android_build/commit/300ffcd78a91e3cb3d5abeadca40b2731b5a23a8

(One thing to note is that this init.rc was setting the vendor and product IDs to Samsung rather than the default Google nexus identification, so if you don't have the proper udev rules for Nexus devices you'll need to set those up. We discussed having it set the correct vid and pid in recovery and Dees_Troy was indifferent about it but other devices are leaving it as Nexus so it seems better not to deviate from that.)

Patch-set: 1
Reviewer: Gerrit User 1000183 <1000183@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=+2
Label: Verified=+1
Label: SUBM=+1