Update patch set 2
Patch Set 2:

I was thinking to remove the version included in n80xx and put for now the exact copy of the files in smdk4412-common in n80xx overlay folder. I'll include a fix I've commited to smdk4412-common about sysfs path of cabc which was wrong (or maybe is different only for our device). After that we can remove settings not supported (like touchwake) or add something else if needed. This is my idea. What do you think about it?

Patch-set: 2